requirements.txt
autopep8==1.7.0
boto3==1.24.96
botocore==1.27.96
certifi==2022.9.24
charset-normalizer==2.1.1
dnspython==2.2.1
idna==3.4
jmespath==1.0.1
pycodestyle==2.9.1
pymongo==4.3.2
python-dateutil==2.8.2
python-dotenv==0.21.0
python3-nmap==1.6.0
requests==2.28.1
s3transfer==0.6.0
simplejson==3.17.6
six==1.16.0
toml==0.10.2
urllib3==1.26.12
xmltodict==0.13.0
